refactor(Routes & Action creators):  Work under construction 🚧
- User profile components updated to use new routes and header
(Token)
