+++ client/src/App.tsx
@@ -8,7 +8,7 @@ import {ThemeProvider} from "styled-components";
 import light from "./styles/themes/light";
 import CreatePostPage from "./pages/CreatePostPage";
 import {useDispatch, useSelector} from "react-redux";
-import {getUser, setUser} from "./redux/actions/actionCreators";
+import {getUser, setUser} from "./redux/actions/public/generalActions";
 import {getPosts} from "./redux/actions/public/postsActions";
 import LoginPage from "./pages/LoginPage";
 import {PostDetailPage} from "./pages/PostDetail";
