+++ client/src/redux/actions/actionCreators.ts
@@ -6,240 +6,9 @@ import {Action} from "./index";
 
 const token = Storage.get("token");
 
-export const setUserCreateModal = (data: any) => {
-  return {type: ActionType.SET_USER_CREATE_MODAL, payload: data};
-};
 
-export const createUser =
-  (userData: any) => async (dispatch: Dispatch<Action>) => {
-    console.log("Data enviada: ", userData);
-    try {
-      let {data} = await axios.post(`/user/register`, userData);
-      console.log("Respuesta:", data);
 
-      const {token} = data;
-      Storage.set('token', token);
-      // let resCreate = await axios.post(`/user/create`, userData);
-      // console.log(resCreate.data);
 
-      // let resGet = await axios.get(`/user/${userData.email}`);
-      // console.log(resGet.data);
-
-      //console.log("Data recibida: ",data );
-      console.log("CREATE USER", "Se envió la data a la API");
-      return dispatch({
-        type: ActionType.GET_USER,
-        payload: {
-          data: data,
-          modal: {
-            val: true,
-            msg: `The user ${userData.email} was created successfully!`,
-          },
-        },
-      });
-    } catch (error) {
-      console.log(error);
-      return dispatch({
-        type: ActionType.SET_POST_CREATE_MODAL,
-        payload: {
-          val: true,
-          msg: `The user ${userData.email}, couldn't be created!`,
-        },
-      });
-    }
-  };
-export const getUser =
-  (userData?: any) => async (dispatch: Dispatch<Action>) => {
-    //console.log("Data enviada: ", userData);
-    try {
-      let res;
-      if (token) {
-        res = await axios.post(`/user/login`, userData,
-          {
-            headers: {
-              token: token,
-            },
-          }
-        );
-      }
-      else {
-        res = await axios.post(`/user/login`, userData);
-        const {token} = res.data;
-        Storage.set('token', token);
-      }
-      const data = res.data;
-      //console.log("Información actualizada");
-      //console.log("Data recibida: ", data);
-      return dispatch({
-        type: ActionType.GET_USER,
-        payload: {
-          data: data,
-          modal: false,
-        },
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const setUser = (userData: any) => {
-  return {
-    type: ActionType.SET_USER,
-    payload: userData,
-  };
-};
-
-export const setEmail =
-  (email: string) => async (dispatch: Dispatch<Action>) => {
-    return dispatch({
-      type: ActionType.SET_EMAIL,
-      payload: email,
-    });
-  };
-
-export const updateUser =
-  (userData: any, userId: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.put(`/applicant/update/${userId}`, userData);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.UPDATE_USER,
-        payload: userData,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const updateUserExp =
-  (userExp: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.put(`/experience/update/${userExp.id}`, userExp);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.UPDATE_USER_EXP,
-        payload: userExp,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const addUserExp =
-  (userExp: any, userId: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.post(`/experience/create/${userId}`, userExp);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.ADD_USER_EXP,
-        payload: userExp,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const deleteUserExp =
-  (id: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      console.log(id);
-      await axios.delete(`/experience/delete/${id}`);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.DELETE_USER_EXP,
-        payload: id,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const updateUserEducation =
-  (userEducation: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.put(`/education/update/${userEducation.id}`, userEducation);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.UPDATE_USER_EDUCATION,
-        payload: userEducation,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const addUserEducation =
-  (userEducation: any, userId: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.post(`/education/create/${userId}`, userEducation);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.ADD_USER_EDUCATION,
-        payload: userEducation,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const deleteUserEducation =
-  (id: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.delete(`/education/delete/${id}`);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.DELETE_USER_EDUCATION,
-        payload: id,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const updateUserLanguages =
-  (userLanguages: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.put(`/language/update/${userLanguages.id}`, userLanguages);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.UPDATE_USER_LANGUAGES,
-        payload: userLanguages,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const addUserLanguages =
-  (userLanguages: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.post(
-        `/language/create/${userLanguages.applicantId}`,
-        userLanguages
-      );
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.ADD_USER_LANGUAGES,
-        payload: userLanguages,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-export const deleteUserLanguages =
-  (id: any) => async (dispatch: Dispatch<Action>) => {
-    try {
-      await axios.delete(`/language/delete/${id}`);
-      console.log("Información actualizada");
-      return dispatch({
-        type: ActionType.DELETE_USER_LANGUAGES,
-        payload: id,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
 
 export const setUserFollows =
   (compId: number, userId: number) => async (dispatch: Dispatch<Action>) => {
